polish, rename, and make it all prettier
