+++ frontend_vue/src/components/tokens/types.ts
@@ -3,3 +3,8 @@ export type MySQLtokenDataType = {
   token: string;
   auth: string;
 };
+
+export type BaseFormValuesType = {
+  email: string;
+  memo: string;
+};
